requirements.txt
black==21.12b0
certifi==2021.10.8
charset-normalizer==2.0.9
click==8.0.3
idna==3.3
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.4.1
python-dotenv==0.19.2
python-http-client==3.3.4
requests==2.26.0
sendgrid==6.9.3
starkbank-ecdsa==2.0.3
tomli==1.2.3
typing-extensions==4.0.1
urllib3==1.26.7
